perf(client/): :art: improve admin dashboard modularization
