+++ client/src/views/admin/components/AllUsersCharts.jsx
@@ -1,0 +1,17 @@
+import { LineChart, TabPanel, Text } from "@tremor/react";
+
+export default function AllUsersCharts({ data }) {
+  return (
+    <TabPanel>
+      <Text className="text-2xl mt-10">Graficos sobre todos los usuarios</Text>
+      <LineChart
+        className="w-full h-96"
+        data={data}
+        index={data.length > 0 ? Object.keys(data[0])[1] : ""}
+        categories={[data.length > 0 ? Object.keys(data[0])[0] : ""]}
+        colors={["amber"]}
+        allowDecimals={false}
+      />
+    </TabPanel>
+  );
+}
